* Fix log level debug default in debug containers * Move log services to package service from infrastructure
