+++ src/cmp/infrastructure/container/BaseConsentContainer.js
@@ -3,7 +3,7 @@ import IABConsentFactory from '../factory/IABConsentFactory'
 import ChainedVendorListRepository from '../repository/ChainedVendorListRepository'
 import InMemoryVendorListRepository from '../repository/InMemoryVendorListRepository'
 import HttpVendorListRepository from '../repository/HttpVendorListRepository'
-import {Log} from '../Log'
+import {Log} from '../service/log/Log'
 import IABVendorConsentsFactory from '../factory/IABVendorConsentsFactory'
 import ConsentStringVendorConsentsRepository from '../repository/ConsentStringVendorConsentsRepository'
 import GetConsentStatusUseCase from '../../application/services/GetConsentStatusUseCase'
